Removed the old frameworks and added xcframeworks.  Updated webp from 1.1.0 to 1.2.2.
